move permissions check only to whep client
