+++ packages/react-native-whip-whep/packages/android-client/MobileWhepClient/src/main/java/com/mobilewhep/client/ClientBase.kt
@@ -1,11 +1,8 @@
 package com.mobilewhep.client
 
-import android.Manifest
 import android.content.Context
-import android.content.pm.PackageManager
 import android.media.AudioAttributes
 import android.util.Log
-import androidx.core.content.ContextCompat
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -69,15 +66,9 @@ open class ClientBase(
   private var listeners = mutableListOf<ClientBaseListener>()
   var onTrackAdded: (() -> Unit)? = null
 
-  private val REQUIRED_PERMISSIONS =
-    arrayOf(Manifest.permission.CAMERA, Manifest.permission.RECORD_AUDIO)
+
 
   init {
-    if (!hasPermissions(appContext, REQUIRED_PERMISSIONS)) {
-      throw PermissionError.PermissionsNotGrantedError(
-        "Permissions for camera and audio recording have not been granted. Please check your application settings."
-      )
-    }
     val iceServers =
       listOf(
         PeerConnection.IceServer
@@ -257,22 +248,6 @@ open class ClientBase(
         }
       )
     }
-
-  private fun hasPermissions(
-    context: Context,
-    permissions: Array<String>
-  ): Boolean {
-    for (permission in permissions) {
-      if (ContextCompat.checkSelfPermission(
-          context,
-          permission
-        ) != PackageManager.PERMISSION_GRANTED
-      ) {
-        return false
-      }
-    }
-    return true
-  }
 
   override fun onSignalingChange(p0: PeerConnection.SignalingState?) {
     Log.d(CLIENT_TAG, "RTC signaling state changed:: ${p0?.name}")
